backend/requirements.txt
Flask==3.0.2
PyMySQL==1.1.1
python-dotenv==1.0.1
openpyxl==3.1.5
